requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic-settings==2.2.1
PyJWT==2.8.0
cryptography==42.0.8
redis==5.0.4
celery==5.4.0
requests==2.32.3
reportlab==4.2.2
pytest==8.2.2
httpx==0.27.0
